+++ layers/Engine/packages/component-model/src/TypeResolvers/ScalarType/BooleanScalarTypeResolver.php
@@ -2,10 +2,9 @@
 
 declare(strict_types=1);
 
-namespace PoP\Engine\TypeResolvers\ScalarType;
+namespace PoP\ComponentModel\TypeResolvers\ScalarType;
 
 use CastToType;
-use PoP\ComponentModel\TypeResolvers\ScalarType\AbstractScalarTypeResolver;
 use stdClass;
 
 /**
