+++ layers/Engine/packages/component-model/src/TypeResolvers/ScalarType/IntScalarTypeResolver.php
@@ -2,10 +2,9 @@
 
 declare(strict_types=1);
 
-namespace PoP\Engine\TypeResolvers\ScalarType;
+namespace PoP\ComponentModel\TypeResolvers\ScalarType;
 
 use CastToType;
-use PoP\ComponentModel\TypeResolvers\ScalarType\AbstractScalarTypeResolver;
 use stdClass;
 
 /**
@@ -13,16 +12,16 @@ use stdClass;
  *
  * @see https://spec.graphql.org/draft/#sec-Scalars.Built-in-Scalars
  */
-class FloatScalarTypeResolver extends AbstractScalarTypeResolver
+class IntScalarTypeResolver extends AbstractScalarTypeResolver
 {
     public function getTypeName(): string
     {
-        return 'Float';
+        return 'Int';
     }
 
     public function getTypeDescription(): ?string
     {
-        return $this->__('The Float scalar type represents float numbers.', 'component-model');
+        return $this->__('The Int scalar type represents non-fractional signed whole numeric values.', 'component-model');
     }
 
     public function coerceValue(string|int|float|bool|stdClass $inputValue): string|int|float|bool|object
@@ -31,10 +30,10 @@ class FloatScalarTypeResolver extends AbstractScalarTypeResolver
             return $error;
         }
 
-        $castInputValue = CastToType::_float($inputValue);
+        $castInputValue = CastToType::_int($inputValue);
         if ($castInputValue === null) {
             return $this->getError($this->getDefaultErrorMessage($inputValue));
         }
-        return (float) $castInputValue;
+        return (int) $castInputValue;
     }
 }
