+++ layers/Engine/packages/component-model/src/TypeResolvers/ScalarType/StringScalarTypeResolver.php
@@ -2,9 +2,8 @@
 
 declare(strict_types=1);
 
-namespace PoP\Engine\TypeResolvers\ScalarType;
+namespace PoP\ComponentModel\TypeResolvers\ScalarType;
 
-use PoP\ComponentModel\TypeResolvers\ScalarType\AbstractScalarTypeResolver;
 use stdClass;
 
 /**
